Update-2018-6-10

+++ plugins/group/index.js
@@ -1,7 +1,7 @@
 const knex = appRequire('init/knex').knex;
 
 const getGroups = () => {
-  return knex('group').where({}).then(s => s);
+  return knex('group').where({}).orderBy('id').then(s => s);
 };
 
 const getOneGroup = id => {
@@ -11,19 +11,20 @@ const getOneGroup = id => {
   });
 };
 
-const addGroup = (name, comment) => {
+const addGroup = (name, comment, showNotice) => {
   return knex('group').insert({
-    name, comment,
+    name, comment, showNotice,
   });
 };
 
-const editGroup = (id, name, comment) => {
+const editGroup = (id, name, comment, showNotice) => {
   return knex('group').update({
-    name, comment,
+    name, comment, showNotice,
   }).where({ id });
 };
 
 const deleteGroup = async id => {
+  if(id === 0) { return; }
   const users = await knex('user').where({ group: id });
   if(users.length > 0) { return Promise.reject('Can not delete group'); }
   await knex('group').delete().where({ id });
